requirements.txt
fastapi
uvicorn[standard]
openai
openpyxl
python-multipart
python-dotenv
